+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.9.1] - Unreleased
+## [0.9.1] - 2025-01-19
 ### Added
+- New option `listed` for uploads to don't show the upload in the homepage.
 - In *markdown* set the `upload` property to `false` to disable the upload button.
 - `collection.documentName` allows to customize the name of the new documents created in a collection. You can use a pattern (`{title}/index.yml`) or a function.
 - `collection.documentLabel` allows to customize the label (public name) of the documents in a collection.
@@ -19,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   ```
 - Added `label` option to collections, documents and uploads.
 - Allow to define uploads using an options object.
-
 
 ### Deprecated
 - `uploads` property for *file* and *markdown* fields. Use `upload` (in singular).
@@ -542,7 +542,7 @@ First version
 [#31]: https://github.com/lumeland/cms/issues/31
 [#34]: https://github.com/lumeland/cms/issues/34
 
-[0.9.1]: https://github.com/lumeland/cms/compare/v0.9.0...HEAD
+[0.9.1]: https://github.com/lumeland/cms/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/lumeland/cms/compare/v0.8.3...v0.9.0
 [0.8.3]: https://github.com/lumeland/cms/compare/v0.8.2...v0.8.3
 [0.8.2]: https://github.com/lumeland/cms/compare/v0.8.1...v0.8.2